requirements.txt
antlr4-python3-runtime==4.9.3
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
confection==0.0.4
cryptography==40.0.1
cymem==2.0.7
emoji==2.2.0
# en-core-web-md==3.5.0
filelock==3.11.0
huggingface-hub==0.13.4
hydra-core==1.3.2
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
lit==16.0.1
MarkupSafe==2.1.2
mpmath==1.3.0
murmurhash==1.0.9
natsort==8.3.1
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
packaging==23.1
pandas==2.0.0
pathspec==0.11.1
pathy==0.10.1
pip==23.0.1
preshed==3.0.8
protobuf==4.22.3
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
setuptools==67.6.1
six==1.16.0
smart-open==6.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stanza==1.5.0
sympy==1.11.1
thinc==8.1.9
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.28.0
triton==2.0.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wasabi==1.1.1
wheel==0.40.0
